105 code display redesign (#132)
* abstracted add favorite logic and style ouput display

* output styling

* no longer overflows horizontally

* fix: grid responsiveness and style
