+++ src/components/InputDisplay.svelte
@@ -16,15 +16,10 @@
   .input-display {
     display: grid;
     grid-template-rows: 1fr 6fr;
-    /* border: 1px solid black; */
-    /* z-index: -2; */
-    min-width: 350px;
+    /* min-width: 350px; */
     height: 100%;
     overflow-y: auto;
     padding: 0 0.5em;
-  }
-  .form-wrapper {
-    /* margin-top: 0px; */
   }
   .info-wrapper {
     margin-top: 2em;
