+++ src/components/icons/StarFilled.svelte
@@ -1,0 +1,1 @@
+<span class="material-icons"> star </span>
