+++ src/components/icons/StarOutline.svelte
@@ -1,0 +1,1 @@
+<span class="material-symbols-outlined"> star </span>
