+++ src/pages/HomePage.svelte
@@ -18,16 +18,17 @@
 <style>
   div {
     display: grid;
-    grid-template-columns: minmax(300px, 1fr) 2fr 2fr;
+    grid-template-columns: minmax(200px, 1fr) 2fr 2fr;
     grid-template-rows: 65px 1fr;
     grid-template-areas:
       "header   header    header"
       "nav      input     output";
-    width: 100%;
+    max-width: 100%;
     height: 100%;
+    min-width: 0;
   }
 
-  @media screen and (max-width: 920px) {
+  @media screen and (max-width: 1100px) {
     div {
       grid-template-columns: 1fr 1fr;
       grid-template-rows: 65px auto 1fr;
@@ -38,7 +39,7 @@
     }
   }
 
-  @media screen and (max-width: 680px) {
+  @media screen and (max-width: 900px) {
     div {
       grid-template-columns: 1fr;
       grid-template-rows: 65px auto 1fr 1fr;
